Redesign: put all the onus on mimic users.
Let users who want to be mimicked the sole controllers over whose
messages get echoed/replaced. Remove "victim" commands and the need for
users to start a conversation with the bot in order to be relayed.

MAKES BREAKING CHANGES TO THE DATABASE. If using a database from an
older version, at least drop the "rooms" and "control_rooms" tables.